# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi",
        "description": "Laravel OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "local server"
        },
        {
            "url": "https://example.com/",
            "description": "staging server"
        },
        {
            "url": "https://api.example.com/",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtenir tots els usuaris",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Llista d'usuaris obtinguda correctament"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear un nou usuari",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "description": "Dades del nou usuari",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuari creat correctament"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtenir un usuari específic",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'usuari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari obtingut correctament"
                    },
                    "404": {
                        "description": "Usuari no trobat"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualitzar un usuari existent",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'usuari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dades actualitzades de l'usuari",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joana"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joana@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuari actualitzat correctament"
                    },
                    "404": {
                        "description": "Usuari no trobat"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar un usuari",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'usuari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuari eliminat correctament"
                    },
                    "404": {
                        "description": "Usuari no trobat"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Autorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}